Moving query_range to tripperware package
