+++ pkg/cortex/modules.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"net/http"
 
+	queryrange2 "github.com/cortexproject/cortex/pkg/querier/tripperware/queryrange"
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/opentracing-contrib/go-stdlib/nethttp"
@@ -32,7 +33,6 @@ import (
 	"github.com/cortexproject/cortex/pkg/frontend/transport"
 	"github.com/cortexproject/cortex/pkg/ingester"
 	"github.com/cortexproject/cortex/pkg/querier"
-	"github.com/cortexproject/cortex/pkg/querier/queryrange"
 	"github.com/cortexproject/cortex/pkg/querier/tenantfederation"
 	"github.com/cortexproject/cortex/pkg/querier/tripperware"
 	querier_worker "github.com/cortexproject/cortex/pkg/querier/worker"
@@ -435,12 +435,12 @@ func (t *Cortex) initDeleteRequestsStore() (serv services.Service, err error) {
 // initQueryFrontendTripperware instantiates the tripperware used by the query frontend
 // to optimize Prometheus query requests.
 func (t *Cortex) initQueryFrontendTripperware() (serv services.Service, err error) {
-	queryRangeMiddlewares, cache, err := queryrange.Middlewares(
+	queryRangeMiddlewares, cache, err := queryrange2.Middlewares(
 		t.Cfg.QueryRange,
 		util_log.Logger,
 		t.Overrides,
-		queryrange.PrometheusCodec,
-		queryrange.PrometheusResponseExtractor{},
+		queryrange2.PrometheusCodec,
+		queryrange2.PrometheusResponseExtractor{},
 		prometheus.DefaultRegisterer,
 		t.TombstonesLoader,
 	)
@@ -453,7 +453,7 @@ func (t *Cortex) initQueryFrontendTripperware() (serv services.Service, err erro
 		prometheus.DefaultRegisterer,
 		t.Cfg.QueryRange.ForwardHeaders,
 		queryRangeMiddlewares,
-		queryrange.PrometheusCodec,
+		queryrange2.PrometheusCodec,
 	)
 
 	return services.NewIdleService(nil, func(_ error) error {
