+++ pkg/querier/tripperware/queryrange/results_cache.go
